Session.js: keep clientSessionId in sync with web storage
The member "clientSessionId" holds the requested client session ID, e.g.
used to support reloading the browser window without loosing the client
session. However, during the startup request, the UI server might choose
to assign a new client session ID. This new ID is stored in the web
storage, but the member "clientSessionId" was not updated.

Change-Id: I06b6d42d43a8153b19d9ac77d17f1361cc8174ec