Set fixed number of parallel jobs for TerserPlugin (Webpack)
When using parallel=true TerserPlugin uses "os.cpus().length - 1" to
compute the number of concurrent build jobs. But os.cpus() returns the
number of cpus of the real hardware in case of a docker container.
As TerserPlugin is used only for production build and production build
usually runs in a docker container environment the parallel option is
set to two processes.