Restore modal views when switching outline 2
Don't avoid bringing the outline to front when the activeForm is modal.
This leads to MessageBoxes/FileChoosers... not to get rendered correctly
when coexisting with a modal view.
Like this, modal views are not automatically brought up front when
switching outlines. But this is the default behavior for non-modal views
as well.

265232

Change-Id: I90bc128c8095ef5698be8a8d30ba407473bd6c86