Webpack: always use external source maps, even in dev mode
The browser is very slow in displaying files with large lines.
If the source map is inlined the last line is very long. This
would not be a problem if the original file were never be opened,
but unfortunately Chrome automatically opens the file some times (e.g.
while debugging).

Change-Id: Icd4f0893e0fa4ec5e302f003464aa38d731c3d56
Reviewed-on: https://git.eclipse.org/r/162831
Tested-by: CI Bot
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>