Fix layouting issues with popups of calendar-component
Use clientY property to define location.y of the popup to ensure
(especially with internet explorer) the popup opens at the height of the
mouse cursor.

Do not reposition the tooltips while layouting desktop if bench-size
did not change to avoid problems during open animations of popups
(compare pixel values to be independent of browser zoom-level).

268925

Change-Id: I1abeb4fef649ecf960b0da0ae29b6a01db596d0f
Reviewed-on: https://git.eclipse.org/r/162891
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/164157
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Adrian Egloff <adrian.egloff@bsi-software.com>