Avoid execChangedValue to be called when TreeBox/ListBox set to null
AbstractTreeBox and AbstractListBox never return null as their value,
even if the value property is null (they return an empty set/list).
This behaviour caused incorrect execChangedValue when setting the
value to null (when it was already null).
With an additional check this wrong call can be avoided.

Added testcase and cleaned up inspections

Signed-off-by: Ralf Muri <ralf.muri@bsi-software.com>
Change-Id: I1aa48d1d4eef4661dff8dfd24397ed5eecee7f1f
Reviewed-on: https://git.eclipse.org/r/162876
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>
Tested-by: CI Bot