Avoid execChangedValue to be called when TreeBox/ListBox set to null
AbstractTreeBox and AbstractListBox never return null as their value,
even if the value property is null (they return an empty set/list).
This behaviour caused incorrect execChangedValue when setting the
value to null (when it was already null).
With an additional check this wrong call can be avoided.

Adds missing changes of previous commit from Ralf