Set preventDoubleClick to true for system buttons and menus
These kind of buttons / menus should typically only be clicked once.
Normally, clicking multiple times should not have a negative effect
beside doing the same thing multiple times (e.g. for search buttons).
But not allowing it could improve performance a little for notorious
double click users and maybe even prevent exception when pressing Ok
twice in some rare cases.

For ScoutJS it should not be a problem because the
form would be closed immediately on the first click anyway, but
it is set nevertheless due to consistency reasons with Scout Classic.

Other buttons and menus are not changed and behave as before because
being able to click multiple times could be desired.

268856

Change-Id: I375afed9e7f8c05116f97adbdd39425aa03f4e33
Reviewed-on: https://git.eclipse.org/r/163318
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>