Webpack: support adding theme entry points and webpack.config.custom.js
Building themes can be expensive especially if there are multiple ones.
Added a function to webpack-defaults.js which generates the entry points
for the themes. This gives the possibility to specify the themes that
should be built using an argument.

Because not every developer may want to work with the same theme,
this configuration should probably not be pushed to the VCS. To make
this easy, the scout-scripts.js now first checks for a file called
webpack.config.custom.js and if available loads this file instead of
webpack.config.js. If this file name is added to .gitignore every
developer can easily customize the configuration without changing
the general config. Example:

webpack.config.custom.js:
const baseConfig = require('./webpack.config');
module.exports = (env, args) => {
  args.themes = args.themes || ['default'];
  return baseConfig(env, args);
};

269936