MenuBarLayout: improve prefSize calculation
If the menu bar contains items which are larger than some others and if
these items are moved to the ellipsis menu when there is not enough
space, the menubar should get smaller. This did not work reliably
because no width hint was passed.

Also, when prefSize is called, menus need to be unshrinked first in
order to get the correct pref size.

Also, calling prefSize must not modify the menubar without resetting it
to the previous state -> added undo param to control it.