MobilePopup: extend from WidgetPopup and add server side implementation
The MobilePopup could only be used on JS side so far, this change
makes it possible to use it from Java code as well.

Also wrapped the setting of animate-open class in a setTimeout function
to give the browser time to layout properly before starting the
animation. With new mobile popups the animations were laggy and didn't
seem to start from the beginning but from the middle (e.g. not from the
bottom of the screen but from the middle).