Form: support type when registering listener, add addUIFormListener
- The Json layer now uses addUIFormListener to handle the events
as late as possible. This fixes the following bug: a form in a popup
is closed. The dispose form event will be sent before the popup close
event -> form is removed before popup is closed which is especially
visible during a close animation (popup should be closed before
the form is removed).

- The new method addFormListener(listener, types) gives the possibility
to listen for specific event types as it is already possible for
other widgets.