Table: add compact mode
The new compact mode automatically optimizes the table for small
screens. If the property is set to true a new column is inserted
and all other columns made invisible. The new CompactColumn only exists
in the Browser, it is not accessible from the Java code.

The transformation is done by a CompactHandler. By default, the
handler creates html code that will be used for a compact cell
by collecting all the contents (text, icon) of the regular cells of
a table row and stacking them vertically. This behavior can easily
be modified by adjusting or exchanging the compact handler.

The property needs to be set manually, there is no responsive
mechanism which will set it depending on the width or by the mobile
device transformer (at least not yet).

272334