Mobile: show default detail form or outline overview
Until now, the top level form or tile outline overview was never
visible on a mobile device. With this change, this root content
will be embedded into the root node. That is why the root node must
be visible.

Only making the tile outline overview visible was not sufficient because
it was too large and the style was optimized for being displayed
in the desktop bench. Optimized sizes and colors so that it embeds
nicely into the outline.

Also fixed the TileOverviewForm.js which always had the wrong
pref size (height=0).
Removed TileOverviewLayout.js because it can easily be replaced with
RowLayout.

Also made sure a compact tile button can show 2 lines but still looks
good when using useUiHeight (-> padding can not be used).

269797