FormBasedLogin: forwarding to login page does not always work
If the requested url contains subfolders, the forwarding to the
login page does not work.
Reason: the login resources (login.js etc.) are located in the root
folder, but if a sub resource returns the login.html, the login
resources cannot be loaded because they are addressed relatively
(e.g. /subfolder/login.js).

Fixed by remembering the original path on the http session and
redirecting after successful login.

Note: normally (if page reloading is sufficient), no session will be
created during the login process. The new redirect feature requires
a session to store the redirect url, but the session is only used
temporarily and will be invalidated after a successful login. So
compared to today, the existing regular login process has not changed
and behaves as usual, only if the new redirect process is activated
a temporary session is created and the regular session is created a
little bit earlier in the process (right after invalidation).

266091

Change-Id: Ia2302a8fadca00f5e0d332928166f0d9d4f6e509
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/165806
Tested-by: Scout Bot <scout-bot@eclipse.org>
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>