Introduce URI scheme 'binref' to allow referencing BinaryResource
An URI with scheme 'binref' identifies a BinaryResource. The utility
class BinaryRefs provides methods to get the BinaryResource reference by
such a URI. An application must create classes implementing
'IBinaryRefHandler' in order to load a BinaryResource for a URI. A
handler registers itself for a path. If a (prefix-)path of a binref URI
matches the registration path of a handler this handler is called to
load BinaryResource for this URI.



Change-Id: I16b333223fcf6ed0f1a23bf5ecca3e2362376684
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/166605
Tested-by: Ralph Steiner <rst@bsiag.com>
Reviewed-by: Ralph Steiner <rst@bsiag.com>
(cherry picked from commit b9aa4537f30da9ab5fa8220850738a5bb7b4e45b)
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/165663
Tested-by: Scout Bot <scout-bot@eclipse.org>