Mobile: always maximize dialogs
Dialogs are now maximized automatically by the device transformer
so they will use the full screen size. It is not necessary and even
confusing that the background is sometimes still partially visible.

If this behavior is not desired it can be disabled completely or
for a specific form by disabling the transformation.

273938