MobileFormPopup: menubar should be shown on the bottom
Prior to commit 6c54e06e, Form.js set the menubar to bottom for
search forms. It looked like the property searchForm was never set, so
it was removed from Form.js. This was not true because it was set by
JsonSearchForm.
Normally the container (table control, form menu) takes care of setting
the property, this is true for search forms, too. That is why
it works for regular search forms. It only does not work for search
forms embedded in a popup.
-> it is now set by the MobilePopup which makes it work for non search
forms, too.

269799