ViewMenuTab: remember selected view button
When toggling the navigationVisible property on the Desktop, the view
button box gets destroyed and re-created at a different place. To
prevent the currently selected item in the ViewMenuTab to get lost
during that process, we remember the lastly selected menu in a new
property "selectedAsMenu" on the corresponding ViewButton.

Change-Id: I2178e17f82b524e16f856fd9b47fbcb62a01bcd0