WizardProgressField: make it more compact when used on mobile
It does not automatically toggle the mode when the container size
changes because it also reduces the height which would make it
misaligned on desktop or tablet where other components could be on
the left or right side.

Also removed max-width because on Chrome it doesn't work and on Safari
it works but since no overflow attribute is defined the text overlaps
the text of the next step -> no need for it.

274595