SimpleTabArea: fix background color of overflow tab
The overflow tab should have the same background-color as a non
selected simple-tab. Currently the background shines through which
is not a problem with the current theme, but if the header
has a different color or if it contains other elements normally
covered by the tab area it is unexpected.

274456