String field validation: always show validation errors in upper/lower case mode
when typing lower case characters in an isFormatUpper() == true string
field, a possible validation error is not always shown. this is because
of a race condition between the validation error event and an
event which syncs the uppercase formatted text from model to ui.
Because the ui already enforces an uppercase format via css, the model
does not have to send its new uppercase value.
The same statement goes for the isFormatLower() == true and typed
lowercase Characters.

Solution: preventing this events also prevents the possible race
condition and ensures the visibility of any validation errors on the
field.

274950

Signed-off-by: Cyrill Wyss <cyrill.wyss@bsi-software.com>
Change-Id: I44a253a99e6a86b193ea0f499c843471df9ecc5d
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/167440
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Andre Wegmueller <awe@bsiag.com>