FormBasedLogin: redirect to url with special characters does not work
If the url contains special characters the redirecting may fail, e.g.
with Safari on os x.

Example:
http://localhost:8082/help/Einführung.htm
will be redirected to
http://localhost:8082/help/Einf%FChrung.htm

%FC is the char ü with ISO 8859.

To make it work it has to be encoded with UTF-8. This needs to be done
manually because sendRedirect does not encode. Each part of the
url needs to be encoded separately, so that / won't be changed.

Additionally, the query string was missing after redirecting ->
it is now saved along with the path info.

266091