Table: support column headers updated event in compact mode
If a header text is changed that is visible in the compact value
the value needs to be recomputed.

It does not work for Scout JS yet because there is no column headers
updated event. If it should be implemented the question arises whether
we should implement such an event or maybe better add
property change events for a column.

272334

Change-Id: Iee2bcc80ae58cd16e5746691df3fb5ceb599a196
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/168723
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>