Add support for focusin and focusout
Until now the focus event has not been delegated to the server
because it would create a lot of requests (e.g. on every tab press).
It would have been possible to enhance a specific widget with that
capability but it required some work.

With this change it will be possible to listen for focus events
but only if the property desktop.trackFocus is activated. Only then
the focused element will be sent to the server.

Change-Id: I6a31e4147d15005c4c97b87a61b984d32d424ccd
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/168956
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>