BEANS.get for beans of the same class and order should be stable
The BeanHierarchy used to store the beans in a HashSet. When calculating
the resulting beans for a hierarchy, this set was first sorted by order
annotation and then all duplicates are removed only keeping the first
bean for each class.
This lead to the situation that for beans having the same class and
order value the "winning" bean depended on the order within the HashSet
which was not deterministic.

Instead the internal data structure is replaced with a list keeping the
bean insertion order stable. As a second criteria to the sorting with
order annotation value the beans reverse insertion order is added. This
makes the result stable and (for beans with the same class and order)
always the "newest" (the one that was added last) is kept.

Change-Id: I2ff97cc0d0272eada3af264534c5d58e385a160f
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/169514
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Andi Bur <andi.bur@gmail.com>