OWASP: unsafe object deserialization
Introduce a blacklist and whitelist concept for Serialization.
Existing SerializationUtility now uses a DefaultSerializerBlacklist as
well as an empty DefaultSerializerWhitelist (empty for not to break
API).

In addition every individual IObjectSerializer can adapt and use a local
blacklist/whitelist.

The global default blacklist/whitelist can be configured with the
config.properties

- scout.serial.blacklistAppend
- scout.serial.blacklistReplace
- scout.serial.whitelist

277440

Change-Id: Ic32660d5d614509ee2e58d4f1d8aacfd05fb2bf1
Signed-off-by: Ivan Motsch <ivan.motsch@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/169783
Tested-by: Scout Bot <scout-bot@eclipse.org>