Tree.js: remove explicit max-width if there is an icon
This did not work correctly with the compact outline because
at the time the data is rendered the compact class is not set
resulting in a wrong width.
Removed width: auto in breadcrumb mode to allow some content to be
right aligned.

Since IE 9 is not supported anymore it can be replaced with flex box.

Change-Id: I07142f3d84682f8a5ec428b9e95bca48755c9966
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/169842
Tested-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>