OWASP: unsafe object deserialization
Introduce a blacklist and whitelist concept for Serialization.
Existing SerializationUtility now uses a DefaultSerializerBlacklist as
well as an empty DefaultSerializerWhitelist (empty for not to break
API).

In addition every individual IObjectSerializer can adapt and use a local
blacklist/whitelist.

The global default blacklist/whitelist can be configured with the
config.properties

- scout.serial.blacklistAppend
- scout.serial.blacklistReplace
- scout.serial.whitelist

277440

Signed-off-by: Ivan Motsch <ivan.motsch@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/169783
Tested-by: Scout Bot <scout-bot@eclipse.org>
Signed-off-by: Ivan Motsch <ivan.motsch@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/169888
Change-Id: I706d299743ac639f29e8240440737c8e3cfdfeb7
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/169956