MenuBarLayout: reset state correctly when calling prefSize
If prefSize of the TabBoxHeader is called the menuBar will be left in
a wrong state so that the menu items may draw over the tab items.

It did not work correctly if widthHint was 0 because menu items
were not shrinked before and reset did not happen in that case as well.

Also, since the overflow code modifies the visibleMenuItems array,
undoing shrink may not have worked for some items.

The filter function in _minSize has been adjusted because there is no
property called ellisis, the correct name would be ellipsis. But since
ellipsis menus are not stackable anyways it does not need to be
explicitly checked.

Change-Id: Ia588585549b9801fced1c4e6dfe58cadfa7521be
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/170082
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>