Mobile: fix exception when selecting element in outline
If a node is selected after some scrolling an exception occurs
because renderFiller is called by the layout after some nodes
are removed by the filter but the viewport (viewRangeRendered) has
not been updated yet.

Actually renderFiller does not need to be called at that time, but
sizeChanged always returned true because the width is fractional
but size() is rounded. Added renderViewport nevertheless to make
sure it does not fail in another case.

278374