Mobile: active state is not removed sometimes
When selecting a node the active state won't be removed. This means
when going back (selecting the parent), the previously selected
node has still the :active pseudo class.

This seems to happen when the active node is made invisible during
mouse down. This is the case because removeSelection() not only removes
the selected class but also child-of-selected which makes the active
node invisible -> Make sure the active node stays visible.

Also adjusted the prev-selected-node animation a little so that the
active colors is shown a little longer but fades away faster.

The other changes don't have a real effect, but still: showing
was not removed when no animation was used (mobile outline) and there
is no need to add a listener when there is already one added.