Fixed issue with wrong glass-pane locking the UI
When using modal views a situation occurred where a message-box is
open in the foreground, but a glass-pane prevents clicks on the
message-box. This problem was introduced with a previous bugfix, see
commit 10f249a.

To solve the issue we add a glass-pane filter to the desktop, which
ensures the container of the message-box is not covered by a glass-
pane.

274353