Mobile: fix exception when menu visibility changes
In some very rare situations (I still don't know what constellation
is really required), an exception occurs if the visibility of
a menu displayed in the node menubar changes.
The menubar of the group box still contained the menu even though
the menubar was made invisible. The visibility change of the menu
then triggered a reordering of the menus in the menubar of the
group box, which relinked the parent because that menubar was not
rendered which eventually removed the menus.
In the test case the menus were rendered again because the listener
of the node menubar also triggered a reordering and rerendered the
menus again.
The solution is to make sure the menubar of the group box never contains
menu items if it is invisible.

Change-Id: I90abd080741e7a7c5f5ef9b4c51d53d95f2ff103
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/171137
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>