SmtpConnectionPool - Improve exception handling
This change introduces some small changes to the handling of exceptions
in the SmtpConnectionPool. This should lead to more cases in which
connections that an exceptions occurred on are not put back into the
pool in order to prevent the pool from being full of defective
connections but not being aware of it and thus reusing the same
defective connections over and over again.

Change-Id: I924f6c26e3ff0a359c53bbb0a9032aac668d39a8
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/171248
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Stephan Merkli <stephan.merkli@bsi-software.com>