FormField: ensure scrollbar is correct after layout
If form field is resized by an animation (e.g. if it is used in a
tile grid), the scrollbar needs to be correct after the layout.
Reason: the tile grid layouts the tiles (incl. the fields) and then
sets the size to the previous size and animates from there.
Since scrollbar update is delayed, the update will happen during
the animation when the size is not correct anymore.

Actually, we need to discuss whether this delayed update should be
removed completely or at least in the layouts.