Fixed modality for message-boxes
We now rely on the DOM order of the modal elements. The top-most
element (incl. message-boxes) should have no glass-pane, which means
the element is clickable. The glass-pane filter functionality is now
reserved for elements that must be outside of the regular modality-
hierarchy. This can be used e.g. for help-popups. If the message-box
is covered by another element, like a modal dialog, it should not be
clickable.

274353