Mobile: tool form won't be shown when bench is displayed
In outline compact mode, the tool menus are only visible when the
outline is displayed. When the bench is displayed no tool menus
are shown. Selecting a tool menu does not display the form right away,
the user has to close the bench first.
Since tool forms don't depend on the rendering state of the menu itself
they can be shown even if the menu is not.

281647

Change-Id: Ic745772d3414670ffbce8001e8425e9b3be1e432
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/172451
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>