Mobile: fix exception with table controls
Since commit 982cc1da an exception can occur when search table control
is selected.
Reason: the TableControlAdapterMenus were never destroyed (this is not
new) and therefore the selected property change was propagated multiple
times because every time a node is selected a new AdapterMenu was
created. Beside the memory leak it did not have an effect before the
mentioned commit. But with the commit the form threw an already rendered
exception because the popup is now rendered even if the table controls
are not (which is the case for the duplicated controls).

Make sure the controls are destroyed correctly so that only one creates
the popup and renders the form.

Also, don't throw an exception if position() is called if popup is not
rendered. This can now happen on page reload while a popup is open
because rendering is requested during init when parent is not rendered
which means rendering will be delayed (see Popup._openWithoutParent).

281647