GroupBoxResponsiveHandler modifications may affect layout validity
Changes applied by the GroupBoxResponsiveHandler may affect the layout.
E.g. if a label visibility changes for a widget with fillHorizontal =
false. In that case the widget must be layouted anew otherwise strange
display errors may occur.
If the invalidation (mark layout is invalid) is suppressed, the widget
will not be layouted because it still believes to be valid.
Instead only a possibly manually called layout validation (e.g. by
calling revalidate) should be suppressed for performance reasons. This
can be safely done because it is during the layout validation of the
component itself which will trigger the validate of the children later
on anyway.

281096

Change-Id: I1947bb3e1bbcee2338a905f5d4066b0cec5df4ad
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/172867
Tested-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>