Fix disabled color for Safari
Variable currentcolor does not seem to work anymore, use @disabled-color
instead.
Because currentcolor does not work anymore, the fill mode needs
to be set as well if color is overridden. This means it needs to be
set for the read-only and selection style.

Also added @read-only-color variable to facilitate adjustments.

282636