EmptySpace menus not executed on disabled table/tree selection
If a disabled row/node is selected in a table/tree, the empty space
context menus could not be executed even though these menus are not
related to the selection.
This was because the root context menu was disabled as soon as the
selection contained disabled elements.
Now the root context menu is only disabled if the table/tree is
completely disabled. On the other hand if only the selection contains
disabled rows (at least one), the single- and multi-selection menus
are disabled only.

274860

Change-Id: I5fe928ec566852f1aa3b4906652559e1b5753edc
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/174055
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>