JandexInventoryBuilder should be able to handle read-only index files
If the Jandex index file is read-only (e.g. because the Scout app is
deployed into a read-only filesystem like a Docker image) the Jandex-
InventoryBuilder failed because it always tried to acquire an exclusive
write lock.

Now a shared read lock is acquired instead but only if the file exists
and cannot be written. Furthermore the builder will not try to store any
updated or newly created indices if the underlying file is read-only. In
that case the application will run with the newly scanned and up-do-date
index but will not store it for later use. A second launch will
therefore scan the location again to ensure the index is up-to-date.

262874

Change-Id: Ife376077ff6111c35ef05391ca223e5eb930d4fc
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/174130
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>