Support unzipping of filenames encoded in Cp437
Zipping a file in windows or with 7-Zip (without the cu parameter) will
encode the filenames in charset Cp437. If one filename contains a
special character (ex. ä,ö,ü), an unzipping operation must be done with
the same charset or it will fail with an IllegalArgumentException.
As this case may likely happen when zip files from users shall be
unzipped, the default unzipping function has been extended.

If the unzipping function with charset UTF-8 fails with an
IllegalArgumentException, the exception is catched and an unzipping with
charset Cp437 will be performed. If this fails as well or Cp437 is not
supported by the JVM, an IllegalArgumentException is thrown.

283162

Signed-off-by: Cyrill Wyss <cyrill.wyss@bsi-software.com>
Change-Id: I1998cdb0b20eda204664776edf4959ce03e019f6
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/174372
Reviewed-by: Andre Wegmueller <awe@bsiag.com>
Tested-by: Scout Bot <scout-bot@eclipse.org>
(cherry picked from commit a1a12081212e02c7c6de265945f407e89877df08)
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/179026