Planner: fix wrong scales when changing the displayMode
When the 'displayMode' is changed on the server model, the scales might
be rendered incorrectly. Reason: Calling setDisplayMode() will
automatically update the 'viewRange' property. Because this is triggered
via a property change listener, the JsonPlanner adapter will be notified
about 'viewRange' before 'displayMode'. On the UI, _renderViewRange
calls _renderScale, which relies on the displayMode.

Fix: ensure property changes for 'displayMode' are processed before any
other properties.

222829

Change-Id: I8f58ef37e615149904a63e3ed611d4540c0b133d
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/174688
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>