Widget: add reveal method
Brings the widget into view by scrolling the first scrollable parent.

Create ScrollOptions class instead of a boolean parameter
so that the other options can be added in the future more easily.
To add the align option, we should probably create
another class that extends ScrollOptions (e.g. ScrollToOptions
or RevealOptions because scrollTop() does not know that property.

Use ScrollOptions class for scrollToTop as well to allow animated
scrolling.

281181

Change-Id: Ic6024829f620b09ee8cdd66d4efb276342d77524
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/174695
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>