Planner: fix wrong scales when changing the displayMode
When the 'displayMode' is changed on the server model, the scales might
be rendered incorrectly. Reason: Calling setDisplayMode() will
automatically update the 'viewRange' property. Because this is triggered
via a property change listener, the JsonPlanner adapter will be notified
about 'viewRange' before 'displayMode'. On the UI, _renderViewRange
calls _renderScale, which relies on the displayMode.

Fix: ensure property changes for 'displayMode' are processed before any
other properties.

222829

Change-Id: I52ec89e7cbe4f19196d3a94ce6c54704fad90444