Deactivate file drop when the field is disabled
Fields which allow to select a file via drag & drop always allow to drop
a file, even if the field is disabled.
This behavior does not fit into the normal disabled behavior of fields.
For example the FileInput.js does not allow to open a file dialog when
disabled but allows to drop a file. The drop behavior on disabled
fields should be to not accept any file drops.

This change uninstalls the drag & drop handler when the field is set to
disabled and installs it when the field is set to enabled.

286452

Signed-off-by: Cyrill Wyss <cyrill.wyss@bsi-software.com>
Change-Id: I5690860699b6865b7d9f34e6063b8d7abeac6584
Signed-off-by: Cyrill Wyss <cyrill.wyss@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/175133
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>