Image: autoFit does not work with IE and useUiHeight true
It fails because of max-height: 100% and max-width: 100%.
The HtmlComponent reads these values when calculating the preferred
size and adjusts the size accordingly. This won't work when the values
are in percent, so the primitive solution is to use a very large pixel
number instead. This does not work for IE because compared to other
browsers IE returns a calculated value rather than defined percentage
value.

283041

Change-Id: I2a99127f0e9d832653f1454d237539b9a08d2414
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/175096
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>