DragAndDrop.js: add drop validation and rename fileDrop to drop
Even though the current implementation only works with files,
it is actually designed to accept other drop events as well in the
future (e.g. text) -> better name the event drop instead of fileDrop to
be prepared.
Also, there is a property called dropMaximumSize on the form field but
the size was only checked when the upload function was called which is
up to the developer. Checking the size should be the default otherwise
the property is confusing.
Widgets that already do the validation can easily deactivate the
default validation on the DragAndDropHandler.

Also add setters for drop type and drop maximum size.

286452

Change-Id: Ia118153649bd69c2c9746401ecba528b0ae5ab0e
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/175312
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>