Table: fix error when closing cell editor popup on sorted smart column
Use case (Scout JS only):
1. sort a smart column
2. click into the cell to open the editor
3. remove the value by clicking the x icon
4. close the popup by clicking outside of the cell

An exception occurs because it tries to open an editor for an already
destroyed field.
The editor is opened by installRow during the reordering of the rows
due to the cell update.
The cell value must not be updated before the popup is
closed -> wait with the update until popup is closed (pass it as
callback function instead of using promises to reduce flickering).

275340

Change-Id: Iaa9c27a7617931630ac1af200bf02fc3ce9f2d58
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/175428
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>