Outline: add inspector hints to pages again
Some Scout versions ago, the inspector hints were added to the page
so that inspecting the page using browsers dev tool revealed the
model class and class id. This feature broke as the tree node logic
was separated from tree into the class TreeNode without moving
the function responsible for the inspector hints.

[Backport of cfbd0a6cb52fc11e756e958e015162a4c6634db4 to 10.0]

287063