Add sortCode to Cell.js
Use the sortCode on a cell when comparing two rows in a column.
Set the sortCode of a code on the cell to sort a SmartColumn by the order of its CodeType.
Use the SortCode of the cells to sort the datasets of the charts on the ChartTableControl.
Remove sorting by size for the pie chart to be consistent with other charts.
Build codeIndexMap after loading codes to avoid visiting all codes on every getCodeIndex() call.

219146

Change-Id: Id39d956946c73634e4968212669e587f6795de8d
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/175437
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>