LargeMemoryPolicy unloads complete outline
if no outline subtree element is selected and memory load is over 90%
then the LargeMemoryPolicy unloads the complete outline even if the
outline is the current outline.

The fix sets default memory level of LargeMemoryPolicy to 100% - it is
the LARGE memory policy. In addition outline root nodes are never
unloaded.

Signed-off-by: Ivan Motsch <ivan.motsch@bsiag.com>
Change-Id: Ib5cb98e389542df44272885a34ddf6e4daf69d3c
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/175576
Tested-by: Scout Bot <scout-bot@eclipse.org>