Table.js: always setting text after edit can break custom columns
The commit dbdecf03 (fix flickering when editing a cell of a smart
column) introduced a non intended breaking change.
Column.buildCell is called when a row is updated
which happens when the cell text changes.
If a custom column overrides that method
and uses the value to build the cell it could fail
because the value is not updated yet.

-> Revert flickering change (not a real cherry pick from 11.0 to prevent
other side effects, but still extracted method to Column.js).

287569