Table.js: always setting text after edit can break custom columns
The commit dbdecf03 (fix flickering when editing a cell of a smart
column) introduced a non intended breaking change.
Column.buildCell is called when a row is updated
which happens when the cell text changes.
If a custom column overrides that method
and uses the value to build the cell it could fail
because the value is not updated yet.

To fix this, the flickering prevention is now moved
to the smart column because it is the only (built in) column that
resolves the text async.

Also add a check in setCellText to prevent an unnecessary double
execution of updateRow. It should actually belong to Column.js
but unfortunately it breaks the boolean column (mouse click on the
checkbox would not toggle it anymore).

287569